Move command logic to own class
